requirements-dev.txt
# Test dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0

# For testing HTTP requests
requests-mock>=1.10.0
